BTGratio to float

+++ R4_STM_CODE/Core/Inc/hx711.hh
@@ -8,7 +8,7 @@ class HX711{
 	GPIO_TypeDef *Sck_gpio;
 	uint16_t Sck_pin;
 	int32_t OffsetInBits;
-	int32_t BitsToGramRatio;
+	float BitsToGramRatio;
 	uint8_t GAIN;
 
 public:
@@ -17,9 +17,9 @@ public:
 	int32_t getWeigthInGramsWithOffset(uint16_t times = 10);
 	float getWeigthInKilogramsWithOffset(uint16_t times = 10);
 	int32_t getOffsetInBits() const;
-	int32_t getBitsToGramRatio() const;
+	float getBitsToGramRatio() const;
 	void setGain(uint8_t gain = 128);
-	void setBitsToGramRatio(int32_t newBitsToGramRatio);
+	void setBitsToGramRatio(float newBitsToGramRatio);
 	void addToOffset(int32_t DifOffsetInBits);
 	void tare();
 	void initialCalibration(int32_t testLoadInGrams, uint16_t calibrationTimeInMilis = 6666);
